--- src/main/java/io/colyseus/serializer/schema/DataChange.java
@@ -1,12 +1,0 @@
-package io.colyseus.serializer.schema;
-
-public class DataChange {
-    public String field;
-    public Object value;
-    public Object previousValue;
-
-    @Override
-    public String toString() {
-        return field + ": " + previousValue + " --> " + value;
-    }
-}
